% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcQIC.R
\name{CalcQIC}
\alias{CalcQIC}
\title{Calculate Quasi-likelihood under the independence model criterion (QIC)}
\usage{
CalcQIC(mod, details = FALSE)
}
\arguments{
\item{mod}{A fitted clogit() model including a cluster term from the survival package.}

\item{details}{If TRUE, provides all the details of QIC calculation}
}
\value{
Returns a dataframe providing the QIC results.
}
\description{
Takes a fit clogit() model with a cluster term and provides the QIC based on methods of Pan et al. 2001. Written by Mathieu Basille and Jerod Merkle. Last updated January 2022.
}
\examples{
# none
}
