% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawRandSteps_emperical.R
\name{DrawRandSteps_emperical}
\alias{DrawRandSteps_emperical}
\title{Draw random steps for SSF analysis}
\usage{
DrawRandSteps_emperical(
  data = data,
  nr = 10,
  distMax = Inf,
  simult = FALSE,
  id_name = "id",
  date_name = "date",
  x_name = "x",
  y_name = "y",
  withinID = FALSE,
  uniform.angles = FALSE
)
}
\arguments{
\item{data}{A dataframe of points (not sf)}

\item{nr}{Number of random steps you want for each observed step}

\item{distMax}{An optional number to specify if you want to limit the greatest distance (in same dist units as your dist column in your dataframe). Default is Inf.}

\item{simult}{When TRUE, the function will simultaneously draw a random step with a random turning angle. If FALSE, the two parameters will be drawn indepenedent of each other.}

\item{id_name}{A character specifying the name of the column representing animal ID.}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{x_name}{A character specifying the name of the column representing x/easting values (in meters).}

\item{y_name}{A character specifying the name of the column representing y/northing values (in meters).}

\item{withinID}{When TRUE, the random steps will be drawn from the distribution of step lengths and turning angles within the ID in question. If FALSE (the Default), it will draw random steps from the distribution of step lengths and turning angles of all animals in the dataset.}

\item{uniform.angles}{If TRUE, you can just draw random/uniform angles and ignore the relative angles in the data. Default is FALSE.}
}
\value{
Returns a dataframe with each row representing a used or available step. The date column now corresponds to the date of the beginning of the step. The strata column provides the strata for the steps.
}
\description{
Takes a dataframe of locations and generates a specified number of random steps. Random steps will be based on the observed distribution of step lengths and turning angles in the data. Written by Jerod Merkle (modeified from Mathieu Basille). Last updated January 2022.
}
\examples{
# none
}
