% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_telonics_iridium_positions.R
\name{fetch_telonics_iridium_positions}
\alias{fetch_telonics_iridium_positions}
\title{Batch download and process Telonics Iridium GPS collar data}
\usage{
fetch_telonics_iridium_positions(
  username = "username",
  password = "password",
  fldr_out = "C:/Users/lolson10/Documents/SourceCode/TelonicsTemp",
  TDC_path = "C:\\\\Program Files (x86)\\\\Telonics\\\\Data Converter\\\\TDC.exe",
  keep.reports = FALSE,
  start_date = Sys.time() - 25 * 24 * 3600
)
}
\arguments{
\item{username}{A character specifying the telonics username}

\item{password}{A character specifying the telonics password}

\item{fldr_out}{An empty folder for writing temporary files. Must use \\ slashes to denote. And the folder must be empty!}

\item{TDC_path}{Character vector of the file pathway to TDC.exe. Must use \\ slashes and must include the TDC.exe at the end of the string. Cand download the TDC here: https://www.telonics.com/software/tdc.php}

\item{keep.reports}{If FALSE (the default), all temporary files and reports will be removed from your temp folder}

\item{start_date}{a posixct object denoting the earliest data that should be returned. If NULL, then ALL data are returned. Default is Sys.time()-10\emph{24}3600}
}
\value{
Returns an sf point dataframe of the relocations. Collars are denoted by CollarSerialNumber, and the GPS.Fix.Time column represents the fix date/time in UTC.
}
\description{
Written by Jerod Merkle. Last updated August 2023.
}
\examples{
# To come
}
