% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUTMZone.R
\name{FindUTMZone}
\alias{FindUTMZone}
\title{Identify UTM zone from longitude and latitude}
\usage{
FindUTMZone(longitude, latitude)
}
\arguments{
\item{longitude}{a numeric vector of longitude values}

\item{latitude}{a numeric vector of latitude values}
}
\value{
Returns numeric value or vector of values representing the UTM zone of your input data.
}
\description{
Can be a single lat/long value or a vector of lat/long values. Written by Jerod Merkle with some help from the internet. Last updated January 2022.
}
\examples{
# none
}
