% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Points2Lines.R
\name{Points2Lines}
\alias{Points2Lines}
\title{Transform points into lines}
\usage{
Points2Lines(
  data = data,
  date_name = "date",
  id_name = "id",
  byid = TRUE,
  no_cores = 4
)
}
\arguments{
\item{data}{A ordered sf POINT dataframe with a posix column and an animal id column}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{id_name}{A character specifying the name of the column representing animal ID.}

\item{byid}{If TRUE, returns a sf line objective with nrow equal to length(unique(data\link{,id})). If FALSE, returns a sf objective with nrow equal to nrow(data)}

\item{no_cores}{How many processor cores would you like to use? e.g., detectCores()-1}
}
\value{
Returns a linestring sf datafame. If byid=TRUE, nrow will be the number of unique ids. If byid==FALSE, nrow of value will be nrow(data)-length(unique(data\link{,id_name}))
}
\description{
Can either create a line string for each id, or a line for each connected point in the database. Written and updated by Jerod Merkle. Last updated January 2022.
}
\examples{
#To come

}
