% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RarifyFixRate.R
\name{RarifyFixRate}
\alias{RarifyFixRate}
\title{Rarify GPS collar point data}
\usage{
RarifyFixRate(
  data = data,
  burst_name = "burst",
  date_name = "date",
  ideal.fix.interval = 3,
  fix.interval.var = 0.5,
  maximize.pts = TRUE,
  max.cpus = detectCores() - 1
)
}
\arguments{
\item{data}{An ordered dataframe or sf POINT dataframe with a posix column and a burst or animal id column}

\item{burst_name}{A character specifying the name of the column representing burst from CalcBurst, or simply animal id.}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{ideal.fix.interval}{Your ideal fix rate in hours}

\item{fix.interval.var}{Variation (in hours) around your ideal fix interval that you are willing to allow.}

\item{maximize.pts}{Do you want to maxmimize the number of fixes with your idea fix interval? If TRUE, you will also have many fixes that are less than your ideal fix interval, but you'll have more at your ideal fix interval.}

\item{max.cpus}{Maximum number of cpus you want this function to use during the parallelization process. Default is detectCores()-1}
}
\value{
Returns a the same dataframe entered but rarified (less rows) based on parameters.
}
\description{
This function takes a dataset of animal locations and subsets it to the step interval of interest.  It will maximize the number of steps available at the ideal fix rate.  It can deal with complicated collar schedules. Written by Jerod Merkle (with help from Tayler LaSharr). Last updated January 2022.
}
\examples{
# To come
}
